requirements.txt
urllib.request
BeautifulSoup
json
pymongo
python-dotenv